Create README.md
Hi guys... Have a great day everyone!  A message to life is about to say something to life.. You all want to be successful, don't you?  So why talk more, let's go to success... these things are very important, watch till the end... so let's see what the 15 points are...